[ICU-6125] Eliminate a few kilobytes of unused bits in constant data structures. X-Commit-URL: https://ssl.icu-project.org/trac/changeset/23182 Former-commit-id: e0fc04c49a3dea01c143751b018348c6122fb8ad

+++ icu4c/source/i18n/dcfmtsym.cpp
@@ -1,6 +1,6 @@
 /*
 *******************************************************************************
-* Copyright (C) 1997-2007, International Business Machines Corporation and    *
+* Copyright (C) 1997-2008, International Business Machines Corporation and    *
 * others. All Rights Reserved.                                                *
 *******************************************************************************
 *
@@ -224,7 +224,7 @@ DecimalFormatSymbols::initialize(const Locale& loc, UErrorCode& status,
 void
 DecimalFormatSymbols::initialize(const UChar** numberElements, int32_t *numberElementsStrLen, int32_t numberElementsLength)
 {
-    static const int32_t TYPE_MAPPING[][2] = {
+    static const int8_t TYPE_MAPPING[][2] = {
         {kDecimalSeparatorSymbol, 0},
         {kGroupingSeparatorSymbol, 1},
         {kPatternSeparatorSymbol, 2},
